Unify version of library packages
When packages versions are different, it's hard to keep them synced.
For example, we make breaking change in CF.Library.Configuration and bump its major version.
Then we should find all dependent packages and also bump their major version.
Then we should find all their dependent packages and bump their version, and so on.
This is a mess.