Remove IFileSystemFacade interface
This is a double-edged sword: to have small facades in specific projects or to have big reusable facade with methods required by many projects.

For this moment I lean toward the former approach. It's very easy to create a facade for specific needs in the project. It seems better than introducing of another dependency on the library.